test: unit testing rebelService

+++ src/main/java/com/rebels/alliance/usecases/RebelService.java
@@ -5,6 +5,7 @@ import com.rebels.alliance.domains.Item;
 import com.rebels.alliance.domains.Rebel;
 import com.rebels.alliance.exceptions.BusinessValidationException;
 import com.rebels.alliance.gateways.RebelGateway;
+import com.rebels.alliance.usecases.validators.InventoryValidator;
 import com.rebels.alliance.usecases.validators.RebelValidator;
 import lombok.AllArgsConstructor;
 import lombok.val;
@@ -17,6 +18,7 @@ import java.util.List;
 public class RebelService {
     private final RebelGateway rebelGateway;
     private final RebelValidator rebelValidator;
+    private final InventoryValidator inventoryValidator;
     private final InventoryService inventoryService;
 
     public Rebel createRebel(Rebel rebel) {
@@ -62,27 +64,21 @@ public class RebelService {
         return rebelGateway.findByParam(parameter, value);
     }
 
-    public Rebel findOneById(int id){
-        return rebelGateway.findOneById(id);
-    }
-
     public void addItem(Rebel rebel, Item item) {
+        val validationErrors = inventoryValidator.validateItem(item);
+        if (!validationErrors.isEmpty()) {
+            throw new BusinessValidationException(validationErrors);
+        }
         Inventory inventory = inventoryService.addItem(item, rebel.getInventory());
         rebel.setInventory(inventory);
-        //Rebel r = updateRebel(rebel);
-        System.out.println("Add item rebel");
-        //System.out.println(r);
     }
 
     public void removeItem(Rebel rebel, Item item) {
-        System.out.println("rebel antes");
-        System.out.println(rebel);
-        System.out.println("item antes");
-        System.out.println(item);
+        val validationErrors = inventoryValidator.validateItem(item);
+        if (!validationErrors.isEmpty()){
+            throw new BusinessValidationException(validationErrors);
+        }
         Inventory inventory = inventoryService.removeItem(item, rebel.getInventory());
         rebel.setInventory(inventory);
-        //Rebel r = updateRebel(rebel);
-        System.out.println("Remove item rebel");
-        //System.out.println(r);
     }
 }
